overhauled almost all of the application, check description for details
Re did most of the entire application, was having a hard time trying to save, and display the saved tasks, and was overthinking how to do it. Added new and improved saveTasks function and displayTasks function, instead of trying to store an array of objects, I just stored the textvalue of each time id, and then pulled them out individually to display them. will add images during day time hours for readme so that is shows the color change correctly